Use an accessory view instead of an intermediate sheet for OPML export
