+++ Mac/AppDelegate.swift
@@ -57,7 +57,7 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSUserInterfaceValidations, 
 	private var addFeedController: AddFeedController?
 	private var addFolderWindowController: AddFolderWindowController?
 	private var importOPMLController: ImportOPMLWindowController?
-	private var exportOPMLController: ExportOPMLWindowController?
+	private var exportOPMLController: ExportOPMLController?
 	private var keyboardShortcutsWindowController: WebViewWindowController?
 	private var inspectorWindowController: InspectorWindowController?
 	private var crashReportWindowController: CrashReportWindowController? // For testing only
@@ -438,7 +438,7 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSUserInterfaceValidations, 
 			return
 		}
 		
-		exportOPMLController = ExportOPMLWindowController()
+		exportOPMLController = ExportOPMLController()
 		exportOPMLController?.runSheetOnWindow(mainWindowController!.window!)
 	}
