+++ Mac/MainWindow/OPML/ExportOPMLAccessoryView.xib
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="15400" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
+    <dependencies>
+        <deployment identifier="macosx"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="15400"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <customObject id="-2" userLabel="File's Owner" customClass="ExportOPMLAccessoryViewController" customModule="NetNewsWire" customModuleProvider="target">
+            <connections>
+                <outlet property="accountPopUpButton" destination="bK0-BJ-siH" id="jce-mY-Lnu"/>
+                <outlet property="view" destination="c22-O7-iKe" id="F7r-Bu-0DA"/>
+            </connections>
+        </customObject>
+        <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
+        <customObject id="-3" userLabel="Application" customClass="NSObject"/>
+        <customView id="c22-O7-iKe">
+            <rect key="frame" x="0.0" y="0.0" width="481" height="127"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMinY="YES"/>
+            <subviews>
+                <customView translatesAutoresizingMaskIntoConstraints="NO" id="lED-IQ-gZC">
+                    <rect key="frame" x="67" y="20" width="347" height="87"/>
+                    <subviews>
+                        <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="56R-DT-OD3">
+                            <rect key="frame" x="-2" y="3" width="63" height="16"/>
+                            <textFieldCell key="cell" scrollable="YES" lineBreakMode="clipping" sendsActionOnEndEditing="YES" alignment="right" title="Account:" id="qLW-sa-J1x">
+                                <font key="font" metaFont="systemBold"/>
+                                <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
+                                <color key="backgroundColor" name="controlColor" catalog="System" colorSpace="catalog"/>
+                            </textFieldCell>
+                        </textField>
+                        <popUpButton verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="bK0-BJ-siH" userLabel="Account Popup">
+                            <rect key="frame" x="65" y="-3" width="285" height="25"/>
+                            <popUpButtonCell key="cell" type="push" title="Item 1" bezelStyle="rounded" alignment="left" lineBreakMode="truncatingTail" state="on" borderStyle="borderAndBezel" imageScaling="proportionallyDown" inset="2" selectedItem="9GQ-mj-Tcp" id="oI2-rj-A5N">
+                                <behavior key="behavior" lightByBackground="YES" lightByGray="YES"/>
+                                <font key="font" metaFont="menu"/>
+                                <menu key="menu" id="J4i-mi-ts0">
+                                    <items>
+                                        <menuItem title="Item 1" state="on" id="9GQ-mj-Tcp"/>
+                                        <menuItem title="Item 2" id="ngU-na-xcU"/>
+                                        <menuItem title="Item 3" id="20Y-AP-vVO"/>
+                                    </items>
+                                </menu>
+                            </popUpButtonCell>
+                            <connections>
+                                <action selector="accountSelected:" target="-2" id="lQN-hm-7Cz"/>
+                            </connections>
+                        </popUpButton>
+                        <textField verticalHuggingPriority="750" horizontalCompressionResistancePriority="250" translatesAutoresizingMaskIntoConstraints="NO" id="xQI-ze-ZFc">
+                            <rect key="frame" x="-2" y="39" width="351" height="48"/>
+                            <textFieldCell key="cell" selectable="YES" id="q1a-Rk-oL6">
+                                <font key="font" metaFont="system"/>
+                                <string key="title">Choose the account with the subscriptions you’d like to export. Subscriptions are exported in the standard OPML format, which most RSS readers can import.</string>
+                                <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
+                                <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
+                            </textFieldCell>
+                        </textField>
+                    </subviews>
+                    <constraints>
+                        <constraint firstAttribute="trailing" secondItem="xQI-ze-ZFc" secondAttribute="trailing" id="6MA-XB-o9x"/>
+                        <constraint firstItem="xQI-ze-ZFc" firstAttribute="leading" secondItem="lED-IQ-gZC" secondAttribute="leading" id="8x9-3N-8Jg"/>
+                        <constraint firstItem="bK0-BJ-siH" firstAttribute="leading" secondItem="56R-DT-OD3" secondAttribute="trailing" constant="8" symbolic="YES" id="AW6-Od-DFn"/>
+                        <constraint firstItem="56R-DT-OD3" firstAttribute="top" secondItem="xQI-ze-ZFc" secondAttribute="bottom" constant="20" id="MLg-wk-R1I"/>
+                        <constraint firstAttribute="trailing" secondItem="bK0-BJ-siH" secondAttribute="trailing" id="MVl-Gh-JDy"/>
+                        <constraint firstItem="bK0-BJ-siH" firstAttribute="firstBaseline" secondItem="56R-DT-OD3" secondAttribute="firstBaseline" id="Njd-pZ-Q9q"/>
+                        <constraint firstItem="56R-DT-OD3" firstAttribute="leading" secondItem="lED-IQ-gZC" secondAttribute="leading" id="O48-xY-l3w"/>
+                        <constraint firstItem="xQI-ze-ZFc" firstAttribute="top" secondItem="lED-IQ-gZC" secondAttribute="top" id="TYW-C0-9VM"/>
+                        <constraint firstAttribute="bottom" secondItem="bK0-BJ-siH" secondAttribute="bottom" id="Uas-t1-rR1"/>
+                        <constraint firstAttribute="width" constant="347" id="cOZ-Ot-MMd"/>
+                    </constraints>
+                </customView>
+            </subviews>
+            <constraints>
+                <constraint firstItem="lED-IQ-gZC" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="c22-O7-iKe" secondAttribute="leading" constant="20" symbolic="YES" id="0F9-Th-uQH"/>
+                <constraint firstItem="lED-IQ-gZC" firstAttribute="centerX" secondItem="c22-O7-iKe" secondAttribute="centerX" id="3jU-jT-k8m"/>
+                <constraint firstItem="lED-IQ-gZC" firstAttribute="top" secondItem="c22-O7-iKe" secondAttribute="top" constant="20" id="908-Vf-gWh"/>
+                <constraint firstAttribute="trailing" relation="greaterThanOrEqual" secondItem="xQI-ze-ZFc" secondAttribute="trailing" id="9c9-5Z-2j3"/>
+                <constraint firstAttribute="bottom" secondItem="lED-IQ-gZC" secondAttribute="bottom" constant="20" id="ppg-6J-6ti"/>
+            </constraints>
+            <point key="canvasLocation" x="139" y="194"/>
+        </customView>
+    </objects>
+</document>
